DL-13479 split AmendsController
