+++ app/views/pages/amends/amendIP14PensionsTakenBetween.scala.html
@@ -41,7 +41,7 @@
     @errorSummary(amendPensionsTakenBetweenForm.errors)
     <h1 class="govuk-heading-xl">@Messages("pla.ip14PensionsTakenBetween.title")</h1>
 
-    @form(action = routes.AmendsController.submitAmendPensionsTakenBetween) {
+    @form(action = routes.AmendsPensionTakenBetweenController.submitAmendPensionsTakenBetween) {
 
         <ul class="govuk-list govuk-list--bullet">
             <li>@Messages("pla.pensionsTakenBetween.bullet.1")</li>
