+++ conf/app.routes
@@ -80,8 +80,8 @@ GET         /print-protection                                                   
 GET         /amend-protection/pensions-taken-before/:protectionType/:status               @controllers.AmendsPensionTakenBeforeController.amendPensionsTakenBefore(protectionType, status)
 POST        /amend-protection/pensions-taken-before                                       @controllers.AmendsPensionTakenBeforeController.submitAmendPensionsTakenBefore
 
-GET         /amend-protection/pensions-taken-between/:protectionType/:status              @controllers.AmendsController.amendPensionsTakenBetween(protectionType, status)
-POST        /amend-protection/pensions-taken-between                                      @controllers.AmendsController.submitAmendPensionsTakenBetween
+GET         /amend-protection/pensions-taken-between/:protectionType/:status              @controllers.AmendsPensionTakenBetweenController.amendPensionsTakenBetween(protectionType, status)
+POST        /amend-protection/pensions-taken-between                                      @controllers.AmendsPensionTakenBetweenController.submitAmendPensionsTakenBetween
 
 GET         /amend-protection/pensions-worth-before/:protectionType/:status               @controllers.AmendsController.amendPensionsWorthBefore(protectionType, status)
 POST        /amend-protection/pensions-worth-before                                       @controllers.AmendsController.submitAmendPensionsWorthBefore
